+++ packages/snack-sdk/CHANGELOG.md
@@ -6,10 +6,14 @@
 
 ### 🎉 New features
 
+### 🐛 Bug fixes
+
+## 3.1.0 — 2020-12-04
+
+### 🎉 New features
+
 - Add `isFeatureSupported` helper function. ([#10](https://github.com/expo/snack/pull/10) by [@IjzerenHein](https://github.com/IjzerenHein))
 - Add `standardizeDependencies` helper function. ([#11](https://github.com/expo/snack/pull/11) by [@IjzerenHein](https://github.com/IjzerenHein))
-
-### 🐛 Bug fixes
 
 ## 3.0.0 — 2020-12-04